+++ driver-core/CHANGELOG.rst
@@ -2,6 +2,8 @@ CHANGELOG
 =========
 
 * 1.0.3:
+  - [api] The query builder now correctly throw an exception when give a value
+    of type it doesn't know about.
   - [bug] Always take per-query CL, tracing, etc. into account for QueryBuilder
     statements (JAVA-143).